edited

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: hharit <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/02 09:57:40 by hharit            #+#    #+#             */
-/*   Updated: 2021/11/03 12:57:50 by hharit           ###   ########.fr       */
+/*   Updated: 2021/11/03 17:00:34 by hharit           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+size_t ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
