+++ ft_strncmp.c
@@ -6,13 +6,14 @@
 /*   By: hharit <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/02 09:57:51 by hharit            #+#    #+#             */
-/*   Updated: 2021/11/02 09:57:53 by hharit           ###   ########.fr       */
+/*   Updated: 2021/11/03 17:09:55 by hharit           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
